feedback type delete functionality with confirmation

+++ FeedbackHub.Domain/Repositories/Interface/IBaseRepository.cs
@@ -17,7 +17,7 @@ namespace FeedbackHub.Domain.Repositories.Interface
         Task<List<T>> FindByAsync(Expression<Func<T, bool>> predicate);
         Task<List<T>> GetAllAsync();
         IQueryable<T> GetAllIncluding(params Expression<Func<T, object>>[] includeProperties);
-        Task<T> GetByIdAsync(long id);
+        Task<T> GetByIdAsync(int id);
         Task<T> UpdateAsync(T t, object key);
         void Delete(T entity);
         void DeleteRange(List<T> entities);
@@ -25,7 +25,7 @@ namespace FeedbackHub.Domain.Repositories.Interface
         void InsertRange(List<T> entities);
         void Update(T entity);
         List<T> GetAll();
-        T GetById(long id);
+        T GetById(int id);
         IQueryable<T> GetQueryable();
         IQueryable<T> GetQueryableWithNoTracking();
     }
